+++ src/backends/capture/opencv_backend.rs
@@ -170,31 +170,37 @@ impl OpenCvCaptureDevice {
     /// If the frame is failed to be read, this will error.
     #[allow(clippy::cast_sign_loss)]
     pub fn raw_frame_vec(&mut self) -> Result<Cow<'_, [u8]>, NokhwaError> {
+        let frame_fmt = Some(self.camera_format.format());
         if !self.is_stream_open() {
-            return Err(NokhwaError::ReadFrameError(
-                "Stream is not open!".to_string(),
-            ));
+            return Err(NokhwaError::ReadFrameError {
+                message: "Stream is not open!".to_string(),
+                format: frame_fmt,
+            });
         }
 
         let mut frame = Mat::default();
         match self.video_capture.read(&mut frame) {
             Ok(a) => {
                 if !a {
-                    return Err(NokhwaError::ReadFrameError(
-                        "Failed to read frame from videocapture: OpenCV return false, camera disconnected?".to_string(),
-                    ));
+                    return Err(NokhwaError::ReadFrameError {
+                        message: "Failed to read frame from videocapture: OpenCV return false, camera disconnected?".to_string(),
+                        format: frame_fmt,
+                    });
                 }
             }
             Err(why) => {
-                return Err(NokhwaError::ReadFrameError(format!(
-                    "Failed to read frame from videocapture: {}",
-                    why
-                )))
+                return Err(NokhwaError::ReadFrameError {
+                    message: format!("Failed to read frame from videocapture: {}", why),
+                    format: frame_fmt,
+                })
             }
         }
 
         if frame.empty() {
-            return Err(NokhwaError::ReadFrameError("Frame Empty!".to_string()));
+            return Err(NokhwaError::ReadFrameError {
+                message: "Frame Empty!".to_string(),
+                format: frame_fmt,
+            });
         }
 
         match frame.size() {
@@ -206,10 +212,13 @@ impl OpenCvCaptureDevice {
                         let frame_data_vec = match Mat::data_typed::<Vec3b>(&frame) {
                             Ok(v) => v,
                             Err(why) => {
-                                return Err(NokhwaError::ReadFrameError(format!(
-                                    "Failed to convert frame into raw Vec3b: {}",
-                                    why
-                                )))
+                                return Err(NokhwaError::ReadFrameError {
+                                    message: format!(
+                                        "Failed to convert frame into raw Vec3b: {}",
+                                        why
+                                    ),
+                                    format: frame_fmt,
+                                })
                             }
                         };
 
@@ -222,19 +231,24 @@ impl OpenCvCaptureDevice {
 
                         Ok(Cow::from(raw_vec))
                     } else {
-                        Err(NokhwaError::ReadFrameError(
-                            "Failed to read frame from videocapture: not cont".to_string(),
-                        ))
+                        Err(NokhwaError::ReadFrameError {
+                            message: "Failed to read frame from videocapture: not cont".to_string(),
+                            format: frame_fmt,
+                        })
                     };
                 }
-                Err(NokhwaError::ReadFrameError(
-                    "Frame width is less than zero!".to_string(),
-                ))
+                Err(NokhwaError::ReadFrameError {
+                    message: "Frame width is less than zero!".to_string(),
+                    format: frame_fmt,
+                })
             }
-            Err(why) => Err(NokhwaError::ReadFrameError(format!(
-                "Failed to read frame from videocapture: failed to read size: {}",
-                why
-            ))),
+            Err(why) => Err(NokhwaError::ReadFrameError {
+                message: format!(
+                    "Failed to read frame from videocapture: failed to read size: {}",
+                    why
+                ),
+                format: frame_fmt,
+            }),
         }
     }
 
@@ -488,9 +502,11 @@ impl CaptureBackendTrait for OpenCvCaptureDevice {
                         if open {
                             return Ok(());
                         }
-                        Err(NokhwaError::OpenStreamError(
-                            "Stream is not opened after stream open attempt opencv".to_string(),
-                        ))
+                        Err(NokhwaError::OpenStreamError {
+                            message: "Stream is not opened after stream open attempt opencv"
+                                .to_string(),
+                            backend: Some(ApiBackend::OpenCv),
+                        })
                     }
                     Err(why) => Err(NokhwaError::OpenDeviceError(
                         idx.to_string(),
@@ -509,9 +525,10 @@ impl CaptureBackendTrait for OpenCvCaptureDevice {
                 if open {
                     return Ok(());
                 }
-                Err(NokhwaError::OpenStreamError(
-                    "Stream is not opened after stream open attempt opencv".to_string(),
-                ))
+                Err(NokhwaError::OpenStreamError {
+                    message: "Stream is not opened after stream open attempt opencv".to_string(),
+                    backend: Some(ApiBackend::OpenCv),
+                })
             }
             Err(why) => Err(NokhwaError::GetPropertyError {
                 property: "Is Stream Open After Open Stream".to_string(),
@@ -549,7 +566,10 @@ impl CaptureBackendTrait for OpenCvCaptureDevice {
     fn stop_stream(&mut self) -> Result<(), NokhwaError> {
         match self.video_capture.release() {
             Ok(_) => Ok(()),
-            Err(why) => Err(NokhwaError::StreamShutdownError(why.to_string())),
+            Err(why) => Err(NokhwaError::StreamShutdownError {
+                message: why.to_string(),
+                backend: Some(ApiBackend::OpenCv),
+            }),
         }
     }
 }